refactor: move sci clubs repository
