+++ lib/features/student_research_group_tab/repositories/scientific_circles_tab_controller.dart
@@ -1,6 +1,6 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import 'scientific_circles_repository.dart';
+import '../../../shared_repositories/sci_clubs_repository/scientific_circles_repository.dart';
 import 'selected_tag_controller.dart';
 import 'tags_repository.dart';
