+++ lib/shared_repositories/sci_clubs_repository/getScientificCircles.graphql
